requirements.txt
google-cloud-storage
requests
urllib3